+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 
 ## 1.9.12
+* AKS Cluster: Add link_to_kubelet_identity
 * VMSS overprovisioning controls
 
 ## 1.9.11